Docs added, tests moved to ce file
