+++ website/content/api-docs/catalog.mdx
@@ -1132,6 +1132,58 @@ $ curl \
 - `FromWildcard` determines whether the service was associated with the gateway by providing a wildcard specifier
   in the gateway's configuration entry
 
+
+## List Exported Services
+
+-> **1.17.2+:** This API is available in Consul versions 1.17.2 and later.
+
+This endpoint returns all the exported services along with their consumers
+
+| Method | Path                                 | Produces           |
+| ------ | ------------------------------------ | ------------------ |
+| `GET`  | `/catalog/exported-services` | `application/json` |
+
+The table below shows this endpoint's support for
+[blocking queries](/consul/api-docs/features/blocking),
+[consistency modes](/consul/api-docs/features/consistency),
+[agent caching](/consul/api-docs/features/caching), and
+[required ACLs](/consul/api-docs/api-structure#authentication).
+
+| Blocking Queries | Consistency Modes | Agent Caching | ACL Required   |
+| ---------------- | ----------------- | ------------- | -------------- |
+| `YES`            | `all`             | `none`        | `mesh:read or operator:read` |
+
+### Sample Request
+
+```shell-session
+$ curl \
+    http://127.0.0.1:8500/v1/catalog/exported-services
+```
+
+### Sample Responses
+
+```json
+[
+    {
+        "Service": {
+            "Name": "backend"
+        },
+        "ConsumerPeers": [
+            "east",
+        ]
+    },
+    {
+        "Service": {
+            "Name": "frontend"
+        },
+        "ConsumerPeers": [
+            "east",
+            "west"
+        ]
+    }
+]
+```
+
 ## Methods to specify namespace <EnterpriseAlert inline />
 
 Catalog endpoints
